+++ src/ControleFinancas.API/Domain/Services/Interfaces/IUsuarioService.cs
@@ -7,7 +7,7 @@ using ControleFinancas.API.DTO.Usuario;
 
 namespace ControleFinancas.API.Domain.Services.Interfaces
 {
-    public interface IUsuarioService : IService<UsuarioRequestContract, UsuarioResponseContract, long >
+    public interface IUsuarioService : IService<UsuarioRequestContract, UsuarioResponseContract, int>
     {
         Task<UsuarioLoginResponseContract> Autenticar(UsuarioLoginRequestContract usuarioLoginRequest);
         Task<UsuarioResponseContract> Obter(string email);
